requirements.txt
py-clob-client>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
rich>=13.0.0
click>=8.1.0
